fix(PSNR.py): :bug: fixed PSNR calculation to account for colored images

+++ PSNR.py
@@ -2,7 +2,11 @@ from math import log10, sqrt
 import cv2 
 import numpy as np 
   
-def calcPSNR(original, compressed): 
+def calcPSNR(original, compressed):
+    # Convert images to float32 for calculation
+    original = original.astype(np.float32)
+    compressed = compressed.astype(np.float32)
+    
     mse = np.mean((original - compressed) ** 2) 
     if(mse == 0):  
         return 100
